登记活动支持给指定题目批量导入图片记录 (#674)
* 創建模版所需的数据库表增加登记活动模板表

* Revert "創建模版所需的数据库表增加登记活动模板表"

This reverts commit fb0e098ad0a57076aaabc6e69c93fcca4b6a0093.

* 在签到活动中增加发送消息的功能

* Revert "在签到活动中增加发送消息的功能"

This reverts commit e856720411d7138e1deecb8945672a7fd6c25402.

* 单图文附件增加规则限制（制作了一部分）

* 登记活动复制bug

* 单图文附件下载需要验证规则

* 修改检查附件是否能下载的逻辑

* 修改

* 111

* 22

* 33

* 单图文附件下载限制初版

* 优化逻辑

* 修改bug

* 11

* 同步平台首页热点单图文得头图

* 分享日志和阅读日志

* 优化

* 11

* 用户浏览共享页，讨论页，专题页记录阅读次数

* 阅读版（enroll、main、index）ok

* 11

* 第二版2.0

* 修改时间戳

* bug

* 登记活动下载题目中的文件

* 版本1

* 111

* 初版（bug：编码；删除无权限）

* 初版1.0（编码。未）

* bug1

* bug2

* 支持excel 和 几个bug

* 1111

* 修改取值逻辑，跳过编码问题

* 优化1

* 优化2

* 版本2，新的解压方式

* 111

* 优化

+++ controllers/pl/fe/matter/enroll/import.php
@@ -34,7 +34,7 @@ class import extends \pl\fe\matter\base {
 		$countOfCols = 0;
 		for ($i = 0, $ii = count($schemas); $i < $ii; $i++) {
 			$schema = $schemas[$i];
-			if (!in_array($schema->type, ['image', 'file'])) {
+			if (!in_array($schema->type, ['file'])) {
 				$objActiveSheet->setCellValueByColumnAndRow($countOfCols, 1, $schema->title);
 				$countOfCols++;
 			}
@@ -73,7 +73,7 @@ class import extends \pl\fe\matter\base {
 	/**
 	 * 上传文件结束
 	 */
-	public function endUpload_action($app) {
+	public function endUpload_action($app, $oneRecordImgNum = 1) {
 		if (false === ($oUser = $this->accountUser())) {
 			return new \ResponseTimeout();
 		}
@@ -82,12 +82,29 @@ class import extends \pl\fe\matter\base {
 		}
 
 		$file = $this->getPostJson();
+		$type = $file->type;
 
 		$oApp = $this->model('matter\enroll')->byId($app, ['cascaded' => 'N']);
 		$modelFs = $this->model('fs/local', $oApp->siteid, '_resumable');
 		$fileUploaded = 'enroll_' . $oApp->id . '_' . $file->name;
-		$records = $this->_extract($oApp, $modelFs->rootDir . '/' . $fileUploaded);
-		$modelFs->delete($fileUploaded);
+		if ($type === 'application/x-zip-compressed') {
+			$recordImgs = $this->_extractZIP($oApp, $modelFs->rootDir . '/' . $fileUploaded, $modelFs, $file);
+			if ($recordImgs[0] === false) {
+				return new \ResponseError($recordImgs[1]);
+			}
+			$data = $this->_extractImg($oApp, $recordImgs[1], $oneRecordImgNum);
+			if ($data[0] === false) {
+				return new \ResponseError($data[1]);
+			}
+			$records = $data[1];
+			// 删除解压后的文件包
+			$this->deldir($recordImgs[1]->toDir);
+		} else if ($type === 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet') {
+			$records = $this->_extractExcel($oApp, $modelFs->rootDir . '/' . $fileUploaded)->records;
+			$modelFs->delete($fileUploaded);
+		} else {
+			return new \ResponseError('暂不支持此格式文件');
+		}
 
 		$eks = $this->_persist($oApp, $records);
 
@@ -95,8 +112,104 @@ class import extends \pl\fe\matter\base {
 	}
 	/**
 	 * 从文件中提取数据
+	 * int $oneRecordImgNum 每条数据抽取多少个图片
 	 */
-	private function &_extract($oApp, $filename) {
+	private function &_extractImg($oApp, $imgs, $oneRecordImgNum = 1) {
+		$schemas = $oApp->dataSchemas;
+		$schemasByTitle = [];
+		foreach ($schemas as $schema) {
+			$schemasByTitle[$schema->title] = $schema;
+		}
+
+		// 取出excel中的数据
+		if (isset($imgs->{$oApp->title})) {
+			$data = $this->_extractExcel($oApp, $imgs->{$oApp->title}['oUrl']);
+			$records = $data->records;
+		} else {
+			$records = [];
+		}
+
+		// 如果有excel，excel决定的数据条数
+		$recordNumExcel = count($records);
+		// 图片决定的数据条数
+		$imgNum = 0;
+		foreach ($imgs->data as $img) {
+			$count = count($img);
+			if ($count > $imgNum) {
+				$imgNum = $count;
+			}
+		}
+		if ($imgNum < $oneRecordImgNum) {
+			$oneRecordImgNum = 1;
+		}
+		$recordNumImg = ceil($imgNum / $oneRecordImgNum);
+		// 决定数据条数
+		if ($recordNumExcel > $recordNumImg) {
+			$recordNum = $recordNumExcel;
+		} else {
+			$recordNum = $recordNumImg;
+		}
+		/**
+		 * 提取数据
+		 * 将图片转换成base64位后再转存
+		 */
+		$newRecords = [];
+		$modelRec = $this->model('matter\enroll\record');
+		$fsuser = $this->model('fs/user', $oApp->siteid);
+		for ($row = 0; $row < $recordNum; $row++) {
+			if (isset($records[$row])) {
+				$oRecord = $records[$row];
+				$oRecData = $records[$row]->data;
+			} else {
+				$oRecord = new \stdClass;
+				$oRecData = new \stdClass;	
+			}
+			foreach ($imgs->data as $key => &$imgArray) {
+				// 从excle表中取出对应列的名称
+				if (isset($schemasByTitle[$key]) && $schemasByTitle[$key]->type === 'image') {
+					$oSchema = $schemasByTitle[$key];
+					// 转存指定数量的图片
+					$base64Imgs = [];
+					for ($i = 1; $i <= $oneRecordImgNum; $i++) {
+						$img = array_shift($imgArray);
+						if (empty($img)) {
+							continue;
+						}
+						//将图片转成base64位储存
+						$mime_type = getimagesize($img['oUrl'])['mime']; 
+				        $base64_data = base64_encode(file_get_contents($img['oUrl']));
+				        $base64_img = 'data:'.$mime_type.';base64,'.$base64_data;
+				        $newImg = new \stdClass;
+				        $newImg->imgSrc = $base64_img;
+
+						$base64Imgs[] = $newImg;
+					}
+					$treatedValue = [];
+					foreach ($base64Imgs as $base64Img) {
+						$rst = $fsuser->storeImg($base64Img);
+						if (false === $rst[0]) {
+							return $rst;
+						}
+						$treatedValue[] = $rst[1];
+					}
+
+					$treatedValue = implode(',', $treatedValue);
+					$oRecData->{$oSchema->id} = $treatedValue;
+				}
+			}
+
+			$oRecord->data = $oRecData;
+			$newRecords[] = $oRecord;
+		}
+
+		unset($records);
+		$rst = [true, $newRecords];
+		return $rst;
+	}
+	/**
+	 * 从文件中提取数据
+	 */
+	private function &_extractExcel($oApp, $filename) {
 		require_once TMS_APP_DIR . '/lib/PHPExcel.php';
 
 		/**
@@ -123,8 +236,11 @@ class import extends \pl\fe\matter\base {
 		 * 提取数据定义信息
 		 */
 		$schemasByCol = [];
+		// 每列名称
+		$clumnNames = [];
 		for ($col = 0; $col < $highestColumnIndex; $col++) {
 			$colTitle = (string) $objWorksheet->getCellByColumnAndRow($col, 1)->getValue();
+			$clumnNames[$col] = $colTitle;
 			if ($colTitle === '备注') {
 				$schemasByCol[$col] = 'comment';
 			} else if ($colTitle === '标签') {
@@ -221,25 +337,96 @@ class import extends \pl\fe\matter\base {
 			$records[] = $oRecord;
 		}
 
-		return $records;
+		$data = new \stdClass;
+		$data->records = $records;
+		$data->clumnNames = $clumnNames;
+		return $data;
+	}
+	/**
+	 * 从文件中提取数据
+	 */
+	private function &_extractZIP($oApp, $zipfile, $modelFs, $file) {
+		require_once TMS_APP_DIR . '/lib/PHPZip.php';
+
+		$savepath = $modelFs->rootDir . '/enroll_' . $oApp->id . '_importZIP_' . $file->uniqueIdentifier;
+		if(!is_dir($savepath)) {  
+        	mkdir($savepath, 0777, true);//创建目录保存解压内容  
+	    }
+	    // 文件目录
+	    $fileDirectory = new \stdClass;
+        $fileDirectory->toDir = $savepath;
+	    if(file_exists($zipfile)) {
+	        $archive   = new \PHPZip();
+	        $FileInfos  = $archive->GetZipInnerFilesInfo($zipfile);
+	        $failFiles = [];  
+	        $pssFiles = [];
+	        for($i=0; $i<count($FileInfos); $i++) {  
+	        	$fileInfo = $FileInfos[$i];
+	            if($fileInfo['folder'] == 0){  
+	            	$rst = $archive->unZip($zipfile, $savepath, $i, $fileInfo);
+	                if($rst['state'] === true){  
+	                	if ((strpos($fileInfo['filename'], '/')) !== false) {
+	                		$oUrl = $savepath . '/' . $fileInfo['filename'];
+	                		$names = explode('/', $fileInfo['filename']);
+	                		$newFile = [];
+	                		$newFile['title'] = $names[1];
+	                		$newFile['size'] = $fileInfo['size'];
+	                		$newFile['oUrl'] = $oUrl;
+	                		$pssFiles[$names[0]][] = $newFile;
+	                	} else {
+	                		$newFile = [];
+	                		$fileName = $fileInfo['filename'];
+	                		$newFile['title'] = substr($fileName, 0, strrpos($fileName, '.'));
+	                		$newFile['size'] = $fileInfo['size'];
+	                		$newFile['oUrl'] = $savepath . '/' . $fileInfo['filename'];
+	                		$fileDirectory->{$newFile['title']} = $newFile;
+	                	}
+	                }else{  
+	                    $failFiles[] = $fileInfo['filename'];  
+	                }  
+	            }else{
+	            	if(!@is_dir($savepath . '/' . $fileInfo['filename'])){ 
+	            	 	@mkdir($savepath . '/' . $fileInfo['filename'], 0777, true); 
+	            	}   
+	            	$fileName = $this->_iconvConvert($fileInfo['filename']);;
+	            	$pssFiles[substr($fileName, 0, -1)] = [];
+	            }  
+	        }
+	        unlink($zipfile);
+
+        	$fileDirectory->data = $pssFiles;
+        	$fileDirectory->failData = $failFiles;
+	    	$res = [true, $fileDirectory];
+	    	return $res;
+	    } else {
+	    	$res = [false, '压缩文件上传失败'];
+	    	return $res;
+	    }
 	}
 	/**
 	 * 保存数据
 	 */
-	private function _persist($oApp, $records) {
+	private function _persist($oApp, $records, $rid = '') {
 		$current = time();
 		$modelApp = $this->model('matter\enroll');
 		$modelRec = $this->model('matter\enroll\record');
 		$enrollKeys = [];
-
+		if (empty($rid)) {
+			if ($activeRound = $this->model('matter\enroll\round')->getActive($oApp)) {
+				$rid = $activeRound->rid;
+			}
+		}
+		
 		foreach ($records as $oRecord) {
 			$ek = $modelRec->genKey($oApp->siteid, $oApp->id);
 
 			$r = array();
 			$r['aid'] = $oApp->id;
+			$r['rid'] = $rid;
 			$r['siteid'] = $oApp->siteid;
 			$r['enroll_key'] = $ek;
 			$r['enroll_at'] = $current;
+			$r['first_enroll_at'] = $current;
 			$r['nickname'] = isset($oRecord->nickname) ? $oRecord->nickname : '';
 			$r['verified'] = isset($oRecord->verified) ? $oRecord->verified : 'N';
 			$r['comment'] = isset($oRecord->comment) ? $oRecord->comment : '';
@@ -265,6 +452,7 @@ class import extends \pl\fe\matter\base {
 					if (count($v)) {
 						$cd = [
 							'aid' => $oApp->id,
+							'rid' => $rid,
 							'enroll_key' => $ek,
 							'schema_id' => $n,
 							'value' => $v,
@@ -277,4 +465,35 @@ class import extends \pl\fe\matter\base {
 
 		return $enrollKeys;
 	}
+	/**
+	 *
+	 */
+	private function _iconvConvert($str, $encoding = 'UTF-8//IGNORE') {
+		$encode = mb_detect_encoding($str, array('ASCII', 'UTF-8', 'GB2312', 'GBK', 'BIG5'));
+
+		if ($encode && $encode != 'UTF-8') {
+			$str = iconv($encode, $encoding, $str);
+		}
+
+		return $str;
+	}
+	/**
+	 * 删除文件夹
+	 */
+	private function deldir($path) {
+        $path .= '/';
+        $files = scandir($path);
+        foreach($files as $file){
+            if($file !="." && $file !=".."){
+                if(is_dir($path . $file)){
+                    $this->deldir($path . $file . '/');
+                }else{
+                    unlink($path . $file);
+                }
+            }
+        }
+        @rmdir($path);
+
+        return 'ok';
+    }
 }
